# contracts/chatbot-api.openapi.yml
openapi: 3.0.3
info:
  title: My Own Chatbot AI API
  version: 0.1.0
  description: |
    Initial MVP API contract for the local-first chatbot application.

    This contract is the source of truth for the first scaffold and keeps the
    .NET + Orleans backend aligned with the React + Zod frontend.

    Scope intentionally excluded from this first version:
    - authentication and authorization
    - token streaming responses
    - persistence/database storage
    - multi-user collaboration features
servers:
  - url: http://localhost:5050
    description: Local development server
...
paths:
...
  /api/models:
    get:
      tags: [Models]
      summary: List locally available Ollama models
      operationId: listModels
      responses:
        '200':
          description: Models loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListModelsResponse'
        '500':
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
...
    ListModelsResponse:
      type: object
      additionalProperties: false
      required:
        - models
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelSummary'
    ModelSummary:
      type: object
      additionalProperties: false
      required:
        - name
        - displayName
        - isDefault
      properties:
        name:
          type: string
        displayName:
          type: string
        isDefault:
          type: boolean
        description:
          type: string
...
    ApiError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        target:
          type: string
        details:
          type: array
          items:
            type: string